requirements.txt
python-docx==0.8.11
pandas==0.25.1